requirements/requirements.txt
wagtail
wagtailfontawesome
wagtailmenus
wagtail-condensedinlinepanel
django-environ
psycopg2
logzero
